Rename namespace to LoadTest
